Revert "Fix/g UI 1"
