+++ img/Captura de pantalla 2024-07-04 092236.png:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+LastWriterPackageFamilyName=Microsoft.ScreenSketch_8wekyb3d8bbwe
+ZoneId=3
