+++ inta/src/back/calculations.py
@@ -2,10 +2,9 @@ import numpy as np
 import math
 import decimal
 import matplotlib.pyplot as plt
-# from inta.src import config
-from inta.src.front.settings import *
+from src import config
 
-decimal.getcontext().prec = PRECISION
+decimal.getcontext().prec = config.PRECISION
 
 class Calculations:
     def __init__(self, gui_services, *args):
@@ -35,7 +34,7 @@ class Calculations:
                 self._plot_calculation_graphs(inverted_distances, halved_substractions, axis)
                 slope = self._slope_calculation(np.array(halved_substractions).astype(np.float64), np.array(inverted_distances).astype(np.float64))
                 
-                result = (slope / MOMENTUM) * FINAL_MOMENTUM
+                result = (slope / config.MOMENTUM) * config.FINAL_MOMENTUM
                 self.steps.append(f'{axis} axis slope = {slope:.15f}')
                 results.append(result)
                 
@@ -71,11 +70,11 @@ class Calculations:
     
     def _slope_calculation(self, y_axis: np.ndarray, x_axis: np.ndarray) -> float:
         slope, intercept = np.polyfit(x_axis, y_axis, 1)
-        x_list = [round(x, PRECISION) for x in x_axis.tolist()]
-        y_list = [round(y, PRECISION) for y in y_axis.tolist()]
+        x_list = [round(x, config.PRECISION) for x in x_axis.tolist()]
+        y_list = [round(y, config.PRECISION) for y in y_axis.tolist()]
         self.steps.append(
-            f'Slope for {[f"{x:.{PRECISION}f}" for x in x_list]} \n'
-            f'and {[f"{y:.{PRECISION}f}" for y in y_list]} \n'
+            f'Slope for {[f"{x:.{config.PRECISION}f}" for x in x_list]} \n'
+            f'and {[f"{y:.{config.PRECISION}f}" for y in y_list]} \n'
             f'slope = {slope:.15f}'
     )
