--- inta/src/config/config.py
@@ -1,7 +1,0 @@
-CSV_NAMES = ["Xmas","Xmenos","Ymas","Ymenos","Zmas","Zmenos"]
-MOMENTUM = 0.0000002
-FINAL_MOMENTUM = 1000
-PRECISION = 15
-PRIMARY_COLOR = "#E7EFE7"
-SECONDARY_COLOR = "#1F62B1"
-IMAGES = ['src/front/resource/X_axis_graph.png','src/front/resource/Y_axis_graph.png','src/front/resource/Z_axis_graph.png']
