--- inta/src/front/settings.py
@@ -1,27 +1,0 @@
-ROTATE_DEFAULT = 0
-ZOOM_DEFAULT = 0
-FLIP_OPTIONS = ['None', 'X', 'Y', 'Both']
-BLUR_DEFAULT = 0
-CONTRAST_DEFAULT = 0
-EFFECT_OPTION = ['None', 'Emboss', 'Find edges', 'Contour', 'Edge enhance']
-BRIGHTNESS_DEFAULT = 1
-VIBRANCE_DEFAULT = 1
-GRAYSCALE_DEFAULT = False
-INVER_DEFAULT = False
-
-BACKCGROUND_COLOR = '#242424'
-WHITE = '#FFF'
-GREY = 'grey'
-BLUE = '#1f6aa5'
-DARK_GREY = '#4a4a4a'
-CLOSE_RED = '#8a0606'
-
-SLIDER_BG = '#64686b'
-
-CSV_NAMES = ["Xmas","Xmenos","Ymas","Ymenos","Zmas","Zmenos"]
-MOMENTUM = 0.0000002
-FINAL_MOMENTUM = 1000
-PRECISION = 15
-PRIMARY_COLOR = "#E7EFE7"
-SECONDARY_COLOR = "#1F62B1"
-IMAGES = ['src/front/resource/X_axis_graph.png','src/front/resource/Y_axis_graph.png','src/front/resource/Z_axis_graph.png']
