+++ inta/src/front/utils.py
@@ -1,6 +1,6 @@
 import tkinter as tk
-from inta.src.front.panels import *
-from inta.src.front.settings import *
+
+from src import config
 from PIL import Image, ImageTk
 
 
@@ -10,7 +10,7 @@ class Utils():
 
     def create_frame(self, frame_in, side_in, complete = False, scrollable = False):
         frame = tk.Frame(frame_in)
-        frame.configure(background = PRIMARY_COLOR)
+        frame.configure(background = config.PRIMARY_COLOR)
 
         if complete:
             frame.pack(side = side_in, fill =  tk.BOTH, expand = True, pady = 10)
@@ -35,7 +35,7 @@ class Utils():
             button = tk.Button(
                 frame_in,
                 text = text_in,
-                background = SECONDARY_COLOR,
+                background = config.SECONDARY_COLOR,
                 name = name_in,
                 command = lambda: command_in(parameter_in),
             )
@@ -48,7 +48,7 @@ class Utils():
             button = tk.Button(
                 frame_in,
                 text = text_in,
-                background = SECONDARY_COLOR,
+                background = config.SECONDARY_COLOR,
                 name = name_in,
                 command = command_in,
             )
@@ -68,7 +68,7 @@ class Utils():
             result_label.pack(pady=20)
         else:
             label = tk.Label(frame_in, text = text_in)
-            label.configure(background = PRIMARY_COLOR)
+            label.configure(background = config.PRIMARY_COLOR)
             if side_in == "":
                 label.pack()
             else:
@@ -80,7 +80,8 @@ class Utils():
         # Aqui tendremos que llamar a la clase de check y poner la funcion en ella
         vcmd = (frame_in.register(self._validate_numeric), '%P')
 
-        EntryPanel(frame_in)
+        entry = tk.Entry(frame_in, validate="key", validatecommand=vcmd)
+        entry.pack()
 
     def create_image_canvas(self, frame, image_path = ""):
         if image_path == "":
@@ -110,14 +111,14 @@ class Utils():
     # TODO could be moved to utils folder or another file to shorten code, go case by case
     # TODO '=' spacing
     def create_scroll(self, frame_main):
-        canvas = tk.Canvas(frame_main, bg = PRIMARY_COLOR)
+        canvas = tk.Canvas(frame_main, bg = config.PRIMARY_COLOR)
         canvas.pack(side=tk.LEFT, fill = tk.BOTH, expand = True)
 
         scrollbar_y = tk.Scrollbar(frame_main, orient = tk.VERTICAL, command = canvas.yview)
         scrollbar_y.pack(side = tk.RIGHT, fill = tk.Y)
         canvas.configure(yscrollcommand = scrollbar_y.set)
 
-        inner_frame = tk.Frame(canvas, bg = PRIMARY_COLOR)
+        inner_frame = tk.Frame(canvas, bg = config.PRIMARY_COLOR)
         window = canvas.create_window((0, 0), window = inner_frame, anchor = "n")
 
         def center_frame(event):
